third_party/WebKit/LayoutTests/fast/events/inputevents/inputevent-keyboard.html

<!DOCTYPE html>
<html>
<head>
<title>InputEvent: beforeinput inputType</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<input type="text" id="txt">
<script>
test(function() {
    var lastBeforeInputType = '';
    var lastInputType = '';
    var txt = document.getElementById('txt');
    txt.addEventListener('beforeinput', function(event) {
        assert_true(event instanceof InputEvent);
        assert_false(event.isComposing);
        lastBeforeInputType = event.inputType;
    });
    txt.addEventListener('input', function(event) {
        assert_true(event instanceof InputEvent);
        assert_false(event.isComposing);
        lastInputType = event.inputType;
    });
    if (!window.eventSender) {
        document.write('This test requires eventSender');
    } else {
        var kNoInputEventFired = 'noInputEventFired';
        function testKeyDownInputType(key, modifiers, inputType) {
            lastBeforeInputType = kNoInputEventFired;
            lastInputType = kNoInputEventFired;
            eventSender.keyDown(key, modifiers);
            assert_equals(lastBeforeInputType, inputType, `${modifiers.toString()}+${key} should produce beforeInputType: ${inputType}`);
            assert_equals(lastInputType, inputType, `${modifiers.toString()}+${key} should produce inputType: ${inputType}`);
        }

        txt.focus();
        // Typing
        testKeyDownInputType('a', [], 'insertText');
        testKeyDownInputType('6', [], 'insertText');
        testKeyDownInputType('Backspace', [], 'deleteContent');
        testKeyDownInputType('l', ['shiftKey'], 'insertText');
        testKeyDownInputType('w', ['shiftKey'], 'insertText');

        // Keyboard commands
        var isMacOS = (navigator.userAgent.indexOf('Mac OS X') != -1);
        if (!isMacOS) {
            // MacOS's eventSender does not work on hot keys other than arrows.
            testKeyDownInputType('z', ['ctrlKey'], 'undo');
            testKeyDownInputType('z', ['ctrlKey', 'shiftKey'], 'redo');
        }
        // Move command should not generate input events.
        testKeyDownInputType('ArrowLeft', [], kNoInputEventFired);
        testKeyDownInputType('ArrowLeft', ['shiftKey'], kNoInputEventFired);
        testKeyDownInputType('Home', [], kNoInputEventFired);
    }
}, 'Testing beforeinput inputType');
</script>
</body>
</html>
